CIのエラー対応
